Added Form to recovery password
Because there was not specified, if I must control if user exists and sending E-Mail, so I didn't do it.
In RecoveryPasswordFormFactory class in onSuccess method will be call some model class which will have method with this interface:
public function recoveryPass(ArrayHash $values): void;

Which is sending email.
So here could be onValidate method, which similary check, if email is registered.

If You have any questions please contact me.